Feat: Implement CreateCustomerContainer for customer creation
- Added CreateCustomerContainer component to handle customer creation with a structured form.
- Integrated form handling using react-hook-form and Zod for validation.
- Updated CreateCustomerPage to utilize CreateCustomerContainer, simplifying the page structure.
- Removed deprecated useCustomerActions file to streamline customer management functionality.
