+++ frontend/src/containers/Customers/index.ts
@@ -1,4 +1,9 @@
 import CustomerContentHeader from "./CustomerContentHeader"
 import CustomerTableContainer from "./CustomerTableContainer"
+import CreateCustomerContainer from "./CreateCustomerContainer"
 
-export { CustomerContentHeader, CustomerTableContainer }
+export {
+  CustomerContentHeader,
+  CustomerTableContainer,
+  CreateCustomerContainer,
+}
