--- frontend/src/containers/Customers/useCustomerActions.ts
@@ -1,30 +1,0 @@
-import { useCallback } from "react"
-
-export interface UseCustomerActionsReturn {
-  handleAddCustomer: () => void
-  handleEditCustomer: (id: string) => void
-  handleDeleteCustomer: (id: string) => void
-}
-
-export const useCustomerActions = (): UseCustomerActionsReturn => {
-  const handleAddCustomer = useCallback(() => {
-    console.log("Add customer clicked")
-    // Implement add customer functionality here
-  }, [])
-
-  const handleEditCustomer = useCallback((id: string) => {
-    console.log("Edit customer:", id)
-    // Implement edit customer functionality here
-  }, [])
-
-  const handleDeleteCustomer = useCallback((id: string) => {
-    console.log("Delete customer:", id)
-    // Implement delete customer functionality here (e.g., confirmation dialog)
-  }, [])
-
-  return {
-    handleAddCustomer,
-    handleEditCustomer,
-    handleDeleteCustomer,
-  }
-}
